+++ pkg/resource/events-logger-config/alloy-events-config_test.go
@@ -58,7 +58,7 @@ func TestGenerateAlloyEventsConfig(t *testing.T) {
 				},
 			}
 
-			config, err := generateAlloyEventsConfig(cluster, tc.includeNamespaces, tc.excludeNamespaces, tc.installationName, false)
+			config, err := generateAlloyEventsConfig(cluster, tc.includeNamespaces, tc.excludeNamespaces, tc.installationName, false, "<tempo-url>")
 			if err != nil {
 				t.Fatalf("Failed to generate alloy config: %v", err)
 			}