+++ pkg/resource/events-logger-config/reconciler.go
@@ -31,8 +31,15 @@ func (r *Resource) ReconcileCreate(ctx context.Context, cluster *capi.Cluster, l
 	logger := log.FromContext(ctx)
 	logger.Info("events-logger-config create")
 
+	// TODO: Error only if tracing is enabled on the installation
+	// Retrieve Tempo ingress name
+	tempoURL, err := common.ReadTempoIngressURL(ctx, cluster, r.Client)
+	if err != nil {
+		return ctrl.Result{}, errors.WithStack(err)
+	}
+
 	// Get desired config
-	desiredEventsLoggerConfig, err := generateEventsLoggerConfig(cluster, loggingAgent, r.IncludeNamespaces, r.ExcludeNamespaces, r.Config.InstallationName, r.Config.InsecureCA)
+	desiredEventsLoggerConfig, err := generateEventsLoggerConfig(cluster, loggingAgent, r.IncludeNamespaces, r.ExcludeNamespaces, r.Config.InstallationName, r.Config.InsecureCA, tempoURL)
 	if err != nil {
 		logger.Info("events-logger-config - failed generating events-logger config!", "error", err)
 		return ctrl.Result{}, errors.WithStack(err)